# openarcFlask/services/Users/static/swagger.yaml
openapi: 3.0.0
info:
  title: OPEN-ARC USER APIS
  description: Open-Arc Project Rest API'S USERS Related
  version: 0.1.9
servers:
  - url: http://127.0.0.1:5000
    description: Local server
  - url: http://ec2-3-15-1-100.us-east-2.compute.amazonaws.com:5000
    description: Droplet server
components:
  securitySchemes:
    bearerAuth:            # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT  
paths:
...
  /profile/:
    post:
      security:
        - bearerAuth: []
      summary: Create Member Profile.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                contact:
                  type: string
                address:  
                  type: string  
                drive:
                  type: string
                  description : Yes/No
                city:
                  type: string
                postal_code:
                  type: string
                hourly_rate:
                  type: string
                profile_pic:
                  type: string 
                  description: Base64 string starting with data:image/jpeg;base64
                cv_string:
                  type: string 
                  description: Base64 string of document
                document_name:
                  type: string
                about:
                  type: string
                unavailable_dates:
                  type: array
                  description : ["2021-04-15","2021-04-25"]
      responses: 
        '200':
          description: Profile Added Successfully
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
        '500':
          description: Server Error
    put:
      security:
        - bearerAuth: []
      summary: Update a Member profile.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                contact:
                  type: string
                address:  
                  type: string  
                drive:
                  type: string
                  description : Yes/No
                city:
                  type: string
                postal_code:
                  type: string
                hourly_rate:
                  type: string
                profile_pic:
                  type: string 
                  description: Base64 string starting with data:image/jpeg;base64
                cv_string:
                  type: string 
                  description: Base64 string of document
                document_name:
                  type: string
                about:
                  type: string
                unavailable_dates:
                  type: array
                  description : ["2021-04-15","2021-04-25"]
                badge_number:
                  type: string
      responses: 
        '200':
          description: Profile Update Successfully
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
        '500':
          description: Server Error
...
  /memberProfile:
    get:
      security:
        - bearerAuth: []
      summary: Get Member profile.
      requestBody:
        required: false
        content:
          application/json:
      responses: 
        '200':
          description: GET Member Profile
...
  /getGuardsList:
    get:
      security:
        - bearerAuth: []
      summary: Get Guards List to Invite for a Job.
      requestBody:
        required: false
        content:
          application/json:
      responses: 
        '200':
          description: Guards Lists
...